+++ Examples/Titanium.Web.Proxy.Examples.Basic/ProxyTestController.cs
@@ -12,10 +12,10 @@ namespace Titanium.Web.Proxy.Examples.Basic
     {
         public void StartProxy()
         {
-            ProxyServer.BeforeRequest += OnRequest;
-            ProxyServer.BeforeResponse += OnResponse;
-            ProxyServer.ServerCertificateValidationCallback += OnCertificateValidation;
-            ProxyServer.ClientCertificateSelectionCallback += OnCertificateSelection;
+            //ProxyServer.BeforeRequest += OnRequest;
+            //ProxyServer.BeforeResponse += OnResponse;
+            //ProxyServer.ServerCertificateValidationCallback += OnCertificateValidation;
+            //ProxyServer.ClientCertificateSelectionCallback += OnCertificateSelection;
 
             //Exclude Https addresses you don't want to proxy
             //Usefull for clients that use certificate pinning
