+++ Titanium.Web.Proxy/Compression/ICompression.cs
@@ -2,6 +2,9 @@
 
 namespace Titanium.Web.Proxy.Compression
 {
+    /// <summary>
+    /// An inteface for http compression
+    /// </summary>
     interface ICompression
     {
         Task<byte[]> Compress(byte[] responseBody);
