+++ Titanium.Web.Proxy/Decompression/DecompressionFactory.cs
@@ -1,5 +1,8 @@
 ﻿namespace Titanium.Web.Proxy.Decompression
 {
+    /// <summary>
+    /// A factory to generate the de-compression methods based on the type of compression
+    /// </summary>
     internal class DecompressionFactory
     {
         internal IDecompression Create(string type)
