+++ Titanium.Web.Proxy/Decompression/DefaultDecompression.cs
@@ -2,6 +2,10 @@
 
 namespace Titanium.Web.Proxy.Decompression
 {
+
+    /// <summary>
+    /// When no compression is specified just return the byte array
+    /// </summary>
     internal class DefaultDecompression : IDecompression
     {
         public Task<byte[]> Decompress(byte[] compressedArray)
