+++ Titanium.Web.Proxy/Decompression/GZipDecompression.cs
@@ -5,6 +5,9 @@ using Titanium.Web.Proxy.Shared;
 
 namespace Titanium.Web.Proxy.Decompression
 {
+    /// <summary>
+    /// concrete implementation of gzip de-compression
+    /// </summary>
     internal class GZipDecompression : IDecompression
     {
         public async Task<byte[]> Decompress(byte[] compressedArray)
@@ -15,7 +18,7 @@ namespace Titanium.Web.Proxy.Decompression
                 using (var output = new MemoryStream())
                 {
                     int read;
-                    while ((read = await decompressor.ReadAsync(buffer, 0, buffer.Length).ConfigureAwait(false)) > 0)
+                    while ((read = await decompressor.ReadAsync(buffer, 0, buffer.Length)) > 0)
                     {
                        await output.WriteAsync(buffer, 0, read);
                     }
