+++ Titanium.Web.Proxy/Decompression/IDecompression.cs
@@ -1,8 +1,10 @@
-﻿using System.IO;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 
 namespace Titanium.Web.Proxy.Decompression
 {
+    /// <summary>
+    /// An interface for decompression
+    /// </summary>
     internal interface IDecompression
     {
        Task<byte[]> Decompress(byte[] compressedArray);
