+++ Titanium.Web.Proxy/EventArguments/CertificateSelectionEventArgs.cs
@@ -3,6 +3,9 @@ using System.Security.Cryptography.X509Certificates;
 
 namespace Titanium.Web.Proxy.EventArguments
 {
+    /// <summary>
+    /// An argument passed on to user for client certificate selection during mutual SSL authentication
+    /// </summary>
     public class CertificateSelectionEventArgs : EventArgs, IDisposable
     {
         public object sender { get; internal set; }
