+++ Titanium.Web.Proxy/Models/EndPoint.cs
@@ -4,6 +4,9 @@ using System.Net.Sockets;
 
 namespace Titanium.Web.Proxy.Models
 {
+    /// <summary>
+    /// An abstract endpoint where the proxy listens
+    /// </summary>
     public abstract class ProxyEndPoint
     {
         public ProxyEndPoint(IPAddress IpAddress, int Port, bool EnableSsl)
@@ -20,6 +23,10 @@ namespace Titanium.Web.Proxy.Models
         internal TcpListener listener { get; set; }
     }
 
+    /// <summary>
+    /// A proxy endpoint that the client is aware of 
+    /// So client application know that it is communicating with a proxy server
+    /// </summary>
     public class ExplicitProxyEndPoint : ProxyEndPoint
     {
         internal bool IsSystemHttpProxy { get; set; }
@@ -34,6 +41,10 @@ namespace Titanium.Web.Proxy.Models
         }
     }
 
+    /// <summary>
+    /// A proxy end point client is not aware of 
+    /// Usefull when requests are redirected to this proxy end point through port forwarding 
+    /// </summary>
     public class TransparentProxyEndPoint : ProxyEndPoint
     {
         //Name of the Certificate need to be sent (same as the hostname we want to proxy)
