+++ Titanium.Web.Proxy/Network/TcpConnectionManager.cs
@@ -10,43 +10,40 @@ using Titanium.Web.Proxy.Helpers;
 using System.Threading;
 using Titanium.Web.Proxy.Extensions;
 using Titanium.Web.Proxy.Shared;
-using System.Security.Cryptography.X509Certificates;
-using Titanium.Web.Proxy.EventArguments;
-using Titanium.Web.Proxy.Models;
 
 namespace Titanium.Web.Proxy.Network
 {
-    public class TcpConnection
-    {
-        internal string HostName { get; set; }
-        internal int port { get; set; }
-        internal bool IsHttps { get; set; }
-        internal Version Version { get; set; }
-
-        internal TcpClient TcpClient { get; set; }
-        internal CustomBinaryReader StreamReader { get; set; }
-        internal Stream Stream { get; set; }
-
-        internal DateTime LastAccess { get; set; }
-
-        internal TcpConnection()
-        {
-            LastAccess = DateTime.Now;
-        }
-    }
-
+    /// <summary>
+    /// A class that manages Tcp Connection to server used by this proxy server
+    /// </summary>
     internal class TcpConnectionManager
     {
+        /// <summary>
+        /// Connection cache
+        /// </summary>
         static Dictionary<string, List<TcpConnection>> connectionCache = new Dictionary<string, List<TcpConnection>>();
+
+        /// <summary>
+        /// A lock to manage concurrency
+        /// </summary>
         static SemaphoreSlim connectionAccessLock = new SemaphoreSlim(1);
 
-
+        /// <summary>
+        /// Get a TcpConnection to the specified host, port optionally HTTPS and a particular HTTP version
+        /// </summary>
+        /// <param name="hostname"></param>
+        /// <param name="port"></param>
+        /// <param name="isHttps"></param>
+        /// <param name="version"></param>
+        /// <returns></returns>
         internal static async Task<TcpConnection> GetClient(string hostname, int port, bool isHttps, Version version)
         {
             List<TcpConnection> cachedConnections = null;
             TcpConnection cached = null;
 
+            //Get a unique string to identify this connection
             var key = GetConnectionKey(hostname, port, isHttps, version);
+
 
             while (true)
             {
@@ -76,6 +73,7 @@ namespace Titanium.Web.Proxy.Network
 
                 if (cached == null)
                     break;
+
             }
 
             if (cached == null)
@@ -90,11 +88,27 @@ namespace Titanium.Web.Proxy.Network
             return cached;
         }
 
+        /// <summary>
+        /// Get a string to identfiy the connection to a particular host, port
+        /// </summary>
+        /// <param name="hostname"></param>
+        /// <param name="port"></param>
+        /// <param name="isHttps"></param>
+        /// <param name="version"></param>
+        /// <returns></returns>
         internal static string GetConnectionKey(string hostname, int port, bool isHttps, Version version)
         {
             return string.Format("{0}:{1}:{2}:{3}:{4}", hostname.ToLower(), port, isHttps, version.Major, version.Minor);
         }
 
+        /// <summary>
+        /// Create connection to a particular host/port optionally with SSL and a particular HTTP version
+        /// </summary>
+        /// <param name="hostname"></param>
+        /// <param name="port"></param>
+        /// <param name="isHttps"></param>
+        /// <param name="version"></param>
+        /// <returns></returns>
         private static async Task<TcpConnection> CreateClient(string hostname, int port, bool isHttps, Version version)
         {
             TcpClient client;
@@ -104,6 +118,7 @@ namespace Titanium.Web.Proxy.Network
             {
                 SslStream sslStream = null;
 
+                //If this proxy uses another external proxy then create a tunnel request for HTTPS connections
                 if (ProxyServer.UpStreamHttpsProxy != null)
                 {
                     client = new TcpClient(ProxyServer.UpStreamHttpsProxy.HostName, ProxyServer.UpStreamHttpsProxy.Port);
@@ -175,7 +190,11 @@ namespace Titanium.Web.Proxy.Network
             };
         }
 
-
+        /// <summary>
+        /// Returns a Tcp Connection back to cache for reuse by other requests
+        /// </summary>
+        /// <param name="connection"></param>
+        /// <returns></returns>
         internal static async Task ReleaseClient(TcpConnection connection)
         {
 
@@ -199,11 +218,17 @@ namespace Titanium.Web.Proxy.Network
 
         private static bool clearConenctions { get; set; }
 
+        /// <summary>
+        /// Stop clearing idle connections
+        /// </summary>
         internal static void StopClearIdleConnections()
         {
             clearConenctions = false;
         }
 
+        /// <summary>
+        /// A method to clear idle connections
+        /// </summary>
         internal async static void ClearIdleConnections()
         {
             clearConenctions = true;
@@ -224,6 +249,7 @@ namespace Titanium.Web.Proxy.Network
                 }
                 finally { connectionAccessLock.Release(); }
 
+                //every minute run this 
                 await Task.Delay(1000 * 60);
             }
